+++ src/Email/Infrastructure/ISmtpClientFactory.cs
@@ -1,8 +1,0 @@
-using MailKit.Net.Smtp;
-
-namespace Aviant.Infrastructure.Email;
-
-public interface ISmtpClientFactory
-{
-    public SmtpClient GetSmtpClient();
-}